python-service/requirements.txt
anthropic==0.28.1
google-genai==1.29.0
httpx>=0.28.1,<1
